# when all builds fail, BDN should stop after printing first error